move generateHexSeed && zeroPadding to utils file
