+++ packages/grid_client/src/primitives/network.ts
@@ -10,8 +10,8 @@ import { RMB } from "../clients/rmb/client";
 import { TFClient } from "../clients/tf-grid/client";
 import { GridClientConfig } from "../config";
 import { events } from "../helpers/events";
-import { formatErrorMessage, getRandomNumber, randomChoice } from "../helpers/utils";
-import { generateRandomHexSeed, validateHexSeed } from "../helpers/validator";
+import { formatErrorMessage, generateRandomHexSeed, getRandomNumber, randomChoice } from "../helpers/utils";
+import { validateHexSeed } from "../helpers/validator";
 import { MyceliumNetworkModel } from "../modules";
 import { DeploymentFactory } from "../primitives/deployment";
 import { BackendStorage, BackendStorageType } from "../storage/backend";
